+++ docs/releases.rst
@@ -9,6 +9,9 @@ v1.3.1 (unreleased)
 New Features
 ~~~~~~~~~~~~
 
+- Adds a Zarr reader to ``open_virtual_dataset``, which allows opening Zarr V2 and V3 stores as virtual datasets.
+  (:pull:`#271`) By `Raphael Hagen <https://github.com/norlandrhagen>`_.
+
 Breaking changes
 ~~~~~~~~~~~~~~~~